renames state class
