+++ app/src/main/java/com/duckduckgo/app/sync/FakeDeviceSyncState.kt
@@ -14,30 +14,21 @@
  * limitations under the License.
  */
 
-package com.duckduckgo.sync.impl
+package com.duckduckgo.app.sync
 
-import com.duckduckgo.appbuildconfig.api.AppBuildConfig
-import com.duckduckgo.appbuildconfig.api.isInternalBuild
 import com.duckduckgo.di.scopes.AppScope
-import com.duckduckgo.sync.api.DeviceSyncRepository
-import com.duckduckgo.sync.api.SyncFeature
+import com.duckduckgo.sync.api.DeviceSyncState
 import com.squareup.anvil.annotations.ContributesBinding
 import javax.inject.*
 
+/**
+ * We need to provide this fake implementation for non internal builds until we can add sync modules dependencies for all flavors.
+ */
 @ContributesBinding(
     scope = AppScope::class,
-    boundType = DeviceSyncRepository::class,
-    priority = ContributesBinding.Priority.HIGHEST,
+    priority = ContributesBinding.Priority.NORMAL,
 )
-class AppDeviceSyncRepository @Inject constructor(
-    private val appBuildConfig: AppBuildConfig,
-    private val syncFeature: SyncFeature,
-    private val syncRepository: SyncRepository,
-) : DeviceSyncRepository {
-
-    override fun isDeviceSyncEnabled(): Boolean = syncRepository.isSignedIn()
-
-    override fun isFeatureEnabled(): Boolean {
-        return syncFeature.self().isEnabled() || appBuildConfig.isInternalBuild()
-    }
+class FakeDeviceSyncState @Inject constructor() : DeviceSyncState {
+    override fun isFeatureEnabled(): Boolean = false
+    override fun isDeviceSyncEnabled(): Boolean = false
 }
